Add base project files
Creates a checkerboard grid of squares as a background with a single
node Entity in the upper-left corner.

The Entity class is the basic unit of drawing and interacting with the
Game. Entities store a reference to their SpriteSheet and can draw onto
the Buffer.

A SpriteSheet provides both a bitmap and data for accessing different
parts of that bitmap, called frames. A collection of usable
SpriteSheets is generated in the SpriteSheetKey class.

Sprite sheet image files are stored in the assets/images/ folder.
Data about sprite sheets is contained inside GameData.json in the
assets/data/ folder.

Other changes to note:
 - Added html-template folder to .gitignore