Add Grid class for storing new Entities
The Grid class has an Entity that it tiles repeatedly as its background
image, and two ways of storing a list of its own Entities. One is a
Vector that simply stores each Entity for drawing purposes, and the
other is an Array that stores all possible positions on the Grid, along
with a reference to what Entity, if any, is stored at each position.
The same Entity will only appear once in the Vector, but can appear
multiple times in the Array, if it is more than 1 grid square in size.

Other changes to note:
 - Added IGameEntity interface for any classes that need to draw
   themselves to the buffer.
 - Deleted TiledEntity. it's behavior is now being duplicated as
   drawRepeatX and drawRepeatY in the Entity class, which determines
   how many times to tile the Entity's image when drawing.