Add dragging to make new Entities
Mouse events in the main class interface with the workbench Grid. The
Grid uses a buffer to track which cells have been visited. Clicking and
releasing in the same cell will create a temporary default wire that
will disappear upon release. To draw a wire, you have to click in one
cell and move to another. For now, all wires are the default look, but
soon the appearance of the wire will depend on the direction in which
you are adding them.