Add the Wire class
Wires are entities that have other Wires as an input and an output. The
presence or absence of these inputs and outputs and the relative
position of the Wires to one another determines how they should be
drawn. All the functionality for choosing what frame for wires to draw
has been moved out of the Grid class and into the Wire class.

Other changes to note:
  - Added the ability to return coordinates expressed in tiles or
    pixels from the getGridCoordinate() function.
  - Removed _dragBuffer and _previousTouch from Grid's variables.
  - Entity.setFrameKey() does not need to be public at the moment.