Add DigitalComponent class
The DigitalComponent class extends Entity and has much the same
functionality of the Wire class previously. The Wire class now extends
the DigitalComponent class and overrides its refresh() function.

Two instances of the DigitalComponent class are added to the starting
Grid as "On" and "Off" constants to play around with.

Updating of disconnected wires is not yet implemented, and it is still
currently possible to create infinite refresh loops if connected Wires
connect back around to themselves.