Add NOT gate and toolbox
The NOT gate is an extension of the DigitalComponent class that inverts
the input given to it.

The upper left corner of the Grid now stores the "toolbox", a
collection of components you can click on to make your active tool.
Clicking on a cell will clone the current tool to that location.
Dragging will still place wires by default.