Refactor and decouple circuit logic
This commit is a partial commit and is part of a major refactoring that
is aimed at separting the circuit logic from the graphical appearance
of entities. After this separation is complete, new functionality will
be added to allow for new Devices (like more logic gates) that can
handle multiple inputs and outputs.

Entities now hold an instance of DigitalComponent, the base class of
all the circuitry. Connectors come in the form of Wires, which connect
to each other or to Nodes, and Nodes, which sit between Wires and
Devices. Devices can be Constants, Gates, or Lamps, for now.

For now, the visuals cannot take into account the direction of each
connecting component, but that will be added back in soon.

Other changes to note:
 - SpriteSheet and SpriteSheetKey were moved out of the entities
   package.
 - The workbench's toolkit has been disabled for now.
 - The Grid's ability to add new components with the mouse has been
   disabled for now.