Restore proper drawing of Connectors
Added tracking of neighbors to Entities so that Connectors can keep
track of which direction they should appear to be going. Entities also
have a drawingLayer property to help determine in what order they
should be drawn.

Note: Drawing of nodes connected to Lamps (a 2x2 entity) will probably
not work as expected, because the direction comparison algorithm is
extremely basic and does not take width or height into account.