Refactor use of the grid
The Grid class has been renamed to Workbench, and a new Grid class has
been added that stores the entities placed onto the Workbench and
controls selection of those entities. Entities now no longer have a
position in pixels, and their position on the grid is used to determine
where they should be drawn on the screen.