Add truth tables
Truth tables use associative arrays (Objects) to map the different
combinations of truth values for the inputs (2 to the power of N, where
N is the number of inputs). Each combination is stored as and must be
searched for with an associative array with keys matching the names of
the inputs. By using the getOutputs() function, you can return the map
for the outputs.

Also, the behavior of Devices has been modified to take advantage of
truth tables for passing values between inputs and outputs. This is the
first step toward being able to easily create new types of logic gates.