Load truth tables from a JSON file
The truth tables used by Devices are now loaded from the same GameData
JSON file that sprite frame keys are loaded from. The format of the
data is illustrated in the Half Adder example below:

"Half Adder" :
{
    "inputs" :
        ["x", "y"],
    "outputs" :
        ["sum", "carry_out"],
    "states" :
        ["0", "1", "1", "2"],
    "outputValues" :
    {
        "sum" :
            [0, 1, 1, 0],
        "carry_out" :
            [0, 0, 0, 1]
    }
}

Values are stored with 0s signifying false and 1s signifying true, and
the ordinal position of the output values following the binary-encoded
position of the inputs. To use the "sum" output above as an example:
Index 0 of the array corresponds to x = false, y = false, and is false.
Index 1 of the array corresponds to x = true,  y = false, and is true.
Index 2 of the array corresponds to x = false, y = true,  and is true.
Index 3 of the array corresponds to x = true,  y = true,  and is false.

Other changes to note:
 - The TruthTable functions clone() and setOutputs() are no longer needed.