Fix bug with truth table outputs
Since the order of inputs can be arbitrary, this was causing erroneous
behavior with truth tables when searching for an output by its input
parameters. To fix this, the index needs to be calculated by the input's
weight instead. Weights are typically powers of 2, to ensure that a
unique index is always possible. If a non-unique index is found, an Error
is thrown.

Searching for an output is now done be recomputing the index and pulling
the output map at that index.