Remove special switch and constant logic
All switches and constants now use the same truth table logic that
other devices use. To accomodate the special interactivity that a
switch provides, states have been added to Devices. Clicking on a
Device with multiple states will cause it to step through the next
state, until reaching the maximum state value where it will reset to
the first state.

To take advantage of this, outputValues and frames in the GameData JSON
file have to list S * (2 ^ I) values in their arrays, where S is the
number of states, and I is the number of inputs. States occupy the
smallest bits, with Inputs occupying the largest bits, so a device with
2 states and 2 inputs would need an array of length 8, with the first
two values corresponding to states 0 and 1 for a false/false input, the
next two values corresponding to states 0 and 1 for a true/false input,
and so on.

Other changes to note:
 - Added a toggleable NOT gate that can be disabled, turning it into a
   buffer, and enabled by clicking on it, like a switch.
 - Added a 4-bit switch that displays its current value in the same
   fashion as a 7-segment display. Clicking on it increments the value
   until it reaches the maximum value, after which it resets to zero.