Use glyphs for fonts and backgrounds
The graphical creation of several devices now uses a tile-based
background and/or text overlay. This is accomplished through the use of
glyphs. In the GameData JSON file, a sprite frame can be declared as
using glyphs by adding three specific elements, "glyphs",
"widthInGlyphs", and "heightInGlyphs". This information is used when
creating the frame keys to generate additional frames based on the size
of glyph and its name.

To use these glyphs, the frames element in an entity must contain three
other elements, "glyphs", "kerning", and "lineHeight". The "glyphs"
element contains a 1- or 2-dimensional array of String values, denoting
the particular that should be used in that position. The "kerning" and
"lineHeight" elements determine the x and y distance between glyphs. If
this value is larger than the width and height of the glyph, there will
be space between glyphs, if this value is smaller, there were be
overlap of glyphs.