Test a NOT gate built using NAND logic
Complex sequential circuits will need to exist as grouped components on
the Board. As a proof of concept, a hard-coded NOT gate built using
NAND logic is being tested here. The internal workings of the NOT gate
are not visible on the Entity layer, but are part of the
DigitalComponent layer.

The hard-coded nature of this will be replaced by a generalized
approach in future commits.